feat: rename directory
