+++ packages/bezier-react/src/components/KeyValueItem/KeyValueItem.const.ts
@@ -1,4 +1,4 @@
-export const TEST_ID_MAP = {
+export const KEY_VALUE_ITEM_TEST_ID = {
   ROOT: 'bezier-react-key-value-list-item',
   MULTILINE_ROOT: 'bezier-react-key-value-list-item-multiline',
   KEY_ITEM: 'bezier-react-key-value-list-item-key-item',
