+++ packages/bezier-react/src/components/KeyValueItem/KeyValueItem.module.scss
@@ -66,7 +66,7 @@
 }
 
 .ValueItem {
-  width: 100%;
+  flex: 1;
 
   /* To avoid inheritance issues with the internal `Text` color */
   color: var(--txt-black-darkest);
