+++ packages/bezier-react/src/components/KeyValueItem/KeyValueItem.stories.tsx
