+++ packages/bezier-react/src/components/KeyValueItem/KeyValueItem.types.ts
