+++ packages/bezier-react/src/index.ts
@@ -48,7 +48,7 @@ export * from '~/src/components/Forms/Slider'
 export * from '~/src/components/Forms/Switch'
 export * from '~/src/components/Help'
 export * from '~/src/components/Icon'
-export * from '~/src/components/KeyValueListItem'
+export * from '~/src/components/KeyValueItem'
 export * from '~/src/components/ListItem'
 export * from '~/src/components/Modal'
 export * from '~/src/components/Navigator/NavGroup'
